refactor: AndroidManifest.xml, Kakao 관련 클래스 정리
